Incorporated the latest working version

+++ icenumerics/vertices.py
@@ -118,7 +118,7 @@ class vertices():
                                 ('id',int),
                                 ('Coordination',int),
                                 ('Charge',int),
-                                ('Dipole',float,(2,))])
+                                ('Dipole',int,(2,))])
         
         self.array['Location'] = v
